requirements.txt
# Base dependencies needed for setup and menu
requests>=2.31.0
rich>=13.7.0
python-dotenv>=1.0.0
setuptools>=69.0.3

# Flask and web dependencies
flask>=3.0.0
werkzeug>=3.0.0
jinja2>=3.1.2
click>=8.1.7
itsdangerous>=2.1.2
flask-sock>=0.7.0
websockets>=11.0.3

# Audio processing and ML dependencies
openai-whisper
sounddevice>=0.4.6
PyAudio>=0.2.14
torch>=2.2.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.1
librosa>=0.10.1
wave>=0.0.2
pydub>=0.25.1

# Utility dependencies
typing_extensions>=4.9.0

# Development dependencies (commented out for production)
# pytest>=7.4.3
# pytest-cov>=4.1.0
# pytest-mock>=3.12.0
# psutil>=5.9.0
# coverage>=7.3.2